online shard split

+++ server/src/main/java/org/opensearch/index/shard/RemoteStoreRefreshListener.java
@@ -22,6 +22,8 @@ import org.opensearch.action.bulk.BackoffPolicy;
 import org.opensearch.action.support.GroupedActionListener;
 import org.opensearch.cluster.routing.RecoverySource;
 import org.opensearch.common.concurrent.GatedCloseable;
+import org.opensearch.common.lease.Releasable;
+import org.opensearch.common.lease.Releasables;
 import org.opensearch.common.logging.Loggers;
 import org.opensearch.common.unit.TimeValue;
 import org.opensearch.common.util.UploadListener;
@@ -46,6 +48,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 import static org.opensearch.index.seqno.SequenceNumbers.LOCAL_CHECKPOINT_KEY;
@@ -87,8 +90,10 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
     private final RemoteSegmentTransferTracker segmentTracker;
     private final Map<String, String> localSegmentChecksumMap;
     private volatile long primaryTerm;
+    private volatile long commitGen = -1;
     private volatile Iterator<TimeValue> backoffDelayIterator;
     private final SegmentReplicationCheckpointPublisher checkpointPublisher;
+    private final AtomicBoolean staleCommitDeletionDelayed = new AtomicBoolean();
 
     public RemoteStoreRefreshListener(
         IndexShard indexShard,
@@ -150,11 +155,24 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
     protected boolean performAfterRefreshWithPermit(boolean didRefresh) {
         boolean successful;
         if (shouldSync(didRefresh, false)) {
-            successful = syncSegments();
+            successful = syncSegments(false);
         } else {
             successful = true;
         }
         return successful;
+    }
+
+    public Releasable delayStaleCommitDeletion() {
+        try (Releasable ignore = drainRefreshes()) {
+            boolean commitsDelayed = staleCommitDeletionDelayed.compareAndSet(false, true);
+            if (commitsDelayed == false) {
+                throw new IllegalStateException("Cannot delay an already delayed stale commit");
+            }
+            return Releasables.releaseOnce(() -> {
+                staleCommitDeletionDelayed.set(false);
+                remoteDirectory.deleteStaleSegmentsAsync(indexShard.getRecoverySettings().getMinRemoteSegmentMetadataFiles());
+            });
+        }
     }
 
     /**
@@ -182,7 +200,7 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
     /*
      @return false if retry is needed
      */
-    private boolean syncSegments() {
+    private boolean syncSegments(boolean firstSyncAfterCommit) {
         if (isReadyForUpload() == false) {
             // Following check is required to enable retry and make sure that we do not lose this refresh event
             // When primary shard is restored from remote store, the recovery happens first followed by changing
@@ -203,12 +221,23 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
                 // if a new segments_N file is present in local that is not uploaded to remote store yet, it
                 // is considered as a first refresh post commit. A cleanup of stale commit files is triggered.
                 // This is done to avoid delete post each refresh.
-                if (isRefreshAfterCommit()) {
+                if (staleCommitDeletionDelayed.get() == false && isRefreshAfterCommit()) {
                     remoteDirectory.deleteStaleSegmentsAsync(indexShard.getRecoverySettings().getMinRemoteSegmentMetadataFiles());
                 }
 
-                try (GatedCloseable<SegmentInfos> segmentInfosGatedCloseable = indexShard.getSegmentInfosSnapshot()) {
+                try (GatedCloseable<SegmentInfos> segmentInfosGatedCloseable = (
+                    firstSyncAfterCommit ?
+                        new GatedCloseable<>(indexShard.store().readLastCommittedSegmentsInfo(), ()->{}) :
+                        indexShard.getSegmentInfosSnapshot()
+                        )
+                ) {
+                    indexShard.store().readLastCommittedSegmentsInfo();
                     SegmentInfos segmentInfos = segmentInfosGatedCloseable.get();
+
+                    if (firstSyncAfterCommit == false && segmentInfos.getGeneration() != commitGen) {
+                        syncSegments(true);
+                        commitGen = segmentInfos.getGeneration();
+                    }
                     final ReplicationCheckpoint checkpoint = indexShard.computeReplicationCheckpoint(segmentInfos);
                     if (checkpoint.getPrimaryTerm() != indexShard.getOperationPrimaryTerm()) {
                         throw new IllegalStateException(
@@ -224,7 +253,6 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
                     // move.
                     long lastRefreshedCheckpoint = ((InternalEngine) indexShard.getEngine()).lastRefreshedCheckpoint();
                     Collection<String> localSegmentsPostRefresh = segmentInfos.files(true);
-
                     // Create a map of file name to size and update the refresh segment tracker
                     updateLocalSizeMapAndTracker(localSegmentsPostRefresh);
                     CountDownLatch latch = new CountDownLatch(1);
@@ -527,6 +555,7 @@ public final class RemoteStoreRefreshListener extends ReleasableRetryableRefresh
         return (indexShard.state() == IndexShardState.RECOVERING && indexShard.shardRouting.primary())
             && indexShard.recoveryState() != null
             && (indexShard.recoveryState().getRecoverySource().getType() == RecoverySource.Type.LOCAL_SHARDS
+                || indexShard.recoveryState().getRecoverySource().getType() == RecoverySource.Type.IN_PLACE_SHARD_SPLIT
                 || indexShard.recoveryState().getRecoverySource().getType() == RecoverySource.Type.SNAPSHOT);
     }
